Updates to use more pytest-bdd features.
