+++ tests/features/generic_functionality.feature
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-@generic_nimbus
+@generic_functionality
 Feature: Generic Functionality smoke tests all pass
 
     @smoke
